# Optimize global identifier access a bit and reduce allocations