… Jint/Runtime/Environments/ObjectEnvironmentRecord.cs
… Jint/Runtime/Interpreter/Expressions/BindingPatternAssignmentExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintArrayExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintBinaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintCallExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintNewExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintObjectExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUpdateExpression.cs
… Jint/Runtime/Interpreter/Statements/JintForStatement.cs
… Jint/Runtime/JavaScriptException.cs
… Jint/Runtime/TypeConverter.cs