Paying some debt from let/const, let see what low-hanging fruit we can find.