@lahma 2020-05-29
## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|634.5 ms|1.67 ms|12000.0000|1000.0000|0.0000|52.64 MB|
| **New** |	|	| **630.1 ms (-1%)** | **1.46 ms** | **12000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **52.62 MB (0%)** |
| Old |Run|3d-morph|531.1 ms|3.00 ms|12000.0000|2000.0000|0.0000|78.1 MB|
| **New** |	|	| **499.1 ms (-6%)** | **2.18 ms** | **12000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **76.44 MB (-2%)** |
| Old |Run|3d-raytrace|447.2 ms|2.15 ms|23000.0000|3000.0000|0.0000|96.09 MB|
| **New** |	|	| **447.1 ms (0%)** | **0.50 ms** | **23000.0000 (0%)** | **3000.0000 (0%)** | **0.0000** | **95.48 MB (-1%)** |
| Old |Run|access-binary-trees|209.9 ms|0.46 ms|20000.0000|1000.0000|0.0000|83.13 MB|
| **New** |	|	| **208.9 ms (0%)** | **0.99 ms** | **20000.0000 (0%)** | **2000.0000 (+100%)** | **0.0000** | **83.08 MB (0%)** |
| Old |Run|access-fannkuch|1,596.6 ms|21.11 ms|9000.0000|0.0000|0.0000|37.95 MB|
| **New** |	|	| **1,550.0 ms (-3%)** | **8.50 ms** | **9000.0000 (0%)** | **0.0000** | **0.0000** | **37.95 MB (0%)** |
| Old |Run|access-nbody|534.4 ms|3.83 ms|15000.0000|0.0000|0.0000|62.82 MB|
| **New** |	|	| **544.5 ms (+2%)** | **2.12 ms** | **15000.0000 (0%)** | **0.0000** | **0.0000** | **62.81 MB (0%)** |
| Old |Run|access-nsieve|596.8 ms|2.37 ms|17000.0000|5000.0000|1000.0000|85.24 MB|
| **New** |	|	| **595.7 ms (0%)** | **0.73 ms** | **17000.0000 (0%)** | **5000.0000 (0%)** | **1000.0000 (0%)** | **85.24 MB (0%)** |
| Old |Run|bitop(...)-byte [24]|394.2 ms|0.95 ms|18000.0000|0.0000|0.0000|75.18 MB|
| **New** |	|	| **399.0 ms (+1%)** | **1.65 ms** | **18000.0000 (0%)** | **0.0000** | **0.0000** | **75.18 MB (0%)** |
| Old |Run|bitops-bits-in-byte|655.1 ms|2.89 ms|12000.0000|0.0000|0.0000|49.87 MB|
| **New** |	|	| **657.1 ms (0%)** | **12.90 ms** | **12000.0000 (0%)** | **0.0000** | **0.0000** | **49.87 MB (0%)** |
| Old |Run|bitops-bitwise-and|642.0 ms|2.45 ms|19000.0000|0.0000|0.0000|77.53 MB|
| **New** |	|	| **422.6 ms (-34%)** | **8.39 ms** | **13000.0000 (-32%)** | **0.0000** | **0.0000** | **54.64 MB (-30%)** |
| Old |Run|bitops-nsieve-bits|598.0 ms|2.50 ms|24000.0000|1000.0000|0.0000|99.28 MB|
| **New** |	|	| **577.7 ms (-3%)** | **11.43 ms** | **24000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **98.9 MB (0%)** |
| Old |Run|contr(...)rsive [21]|309.8 ms|1.70 ms|32000.0000|0.0000|0.0000|128.3 MB|
| **New** |	|	| **276.3 ms (-11%)** | **1.69 ms** | **32000.0000 (0%)** | **0.0000** | **0.0000** | **128.3 MB (0%)** |
| Old |Run|crypto-aes|406.5 ms|1.68 ms|6000.0000|1000.0000|0.0000|28.19 MB|
| **New** |	|	| **385.9 ms (-5%)** | **7.66 ms** | **6000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **28.2 MB (0%)** |
| Old |Run|crypto-md5|268.3 ms|2.23 ms|24000.0000|2000.0000|0.0000|99.17 MB|
| **New** |	|	| **247.3 ms (-8%)** | **4.91 ms** | **24000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **99.15 MB (0%)** |
| Old |Run|crypto-sha1|268.5 ms|1.29 ms|21000.0000|1000.0000|0.0000|86.89 MB|
| **New** |	|	| **255.2 ms (-5%)** | **5.09 ms** | **21000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **86.88 MB (0%)** |
| Old |Run|date-format-tofte|312.1 ms|0.90 ms|17000.0000|1000.0000|0.0000|70.32 MB|
| **New** |	|	| **299.4 ms (-4%)** | **5.95 ms** | **17000.0000 (0%)** | **0.0000 (-100%)** | **0.0000** | **69.7 MB (-1%)** |
| Old |Run|date-format-xparb|154.6 ms|0.52 ms|7000.0000|1000.0000|0.0000|29.76 MB|
| **New** |	|	| **138.3 ms (-11%)** | **2.70 ms** | **7000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **29.61 MB (-1%)** |
| Old |Run|math-cordic|951.5 ms|2.74 ms|31000.0000|0.0000|0.0000|126.66 MB|
| **New** |	|	| **906.9 ms (-5%)** | **17.97 ms** | **31000.0000 (0%)** | **0.0000** | **0.0000** | **125.7 MB (-1%)** |
| Old |Run|math-partial-sums|391.6 ms|1.74 ms|15000.0000|0.0000|0.0000|63.5 MB|
| **New** |	|	| **368.5 ms (-6%)** | **1.22 ms** | **13000.0000 (-13%)** | **0.0000** | **0.0000** | **53.81 MB (-15%)** |
| Old |Run|math-spectral-norm|308.7 ms|1.03 ms|18000.0000|0.0000|0.0000|72.29 MB|
| **New** |	|	| **308.6 ms (0%)** | **0.62 ms** | **18000.0000 (0%)** | **0.0000** | **0.0000** | **72.28 MB (0%)** |
| Old |Run|regexp-dna|434.6 ms|0.46 ms|0.0000|0.0000|0.0000|18.31 MB|
| **New** |	|	| **442.0 ms (+2%)** | **1.91 ms** | **0.0000** | **0.0000** | **0.0000** | **18.3 MB (0%)** |
| Old |Run|string-base64|239.1 ms|0.85 ms|3000.0000|1000.0000|0.0000|13.76 MB|
| **New** |	|	| **231.3 ms (-3%)** | **0.63 ms** | **3000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **13.13 MB (-5%)** |
| Old |Run|string-fasta|556.9 ms|2.55 ms|41000.0000|0.0000|0.0000|165.85 MB|
| **New** |	|	| **533.2 ms (-4%)** | **1.54 ms** | **41000.0000 (0%)** | **0.0000** | **0.0000** | **164.92 MB (-1%)** |
| Old |Run|string-tagcloud|194.3 ms|3.79 ms|11000.0000|1000.0000|0.0000|56 MB|
| **New** |	|	| **199.1 ms (+2%)** | **2.59 ms** | **11000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **56 MB (0%)** |
| Old |Run|string-unpack-code|205.0 ms|1.91 ms|20000.0000|3000.0000|1000.0000|89.33 MB|
| **New** |	|	| **206.3 ms (+1%)** | **0.37 ms** | **20000.0000 (0%)** | **3000.0000 (0%)** | **1000.0000 (0%)** | **89.33 MB (0%)** |
| Old |Run|strin(...)input [21]|1,277.9 ms|12.42 ms|1126000.0000|1105000.0000|1104000.0000|6211.79 MB|
| **New** |	|	| **1,262.1 ms (-1%)** | **9.60 ms** | **1170000.0000 (+4%)** | **1150000.0000 (+4%)** | **1149000.0000 (+4%)** | **6211.09 MB (0%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Slice|100|551.2 μs|2.19 μs|181.6406|0.0000|0.0000|742.19 KB|
| **New** |	|	| **543.9 μs (-1%)** | **1.61 μs** | **180.6641 (-1%)** | **0.0000** | **0.0000** | **739.84 KB (0%)** |
| Old |Concat|100|750.2 μs|4.32 μs|216.7969|0.0000|0.0000|886.72 KB|
| **New** |	|	| **745.5 μs (-1%)** | **3.14 μs** | **215.8203 (0%)** | **0.0000** | **0.0000** | **884.38 KB (0%)** |
| Old |Unshift|100|21,908.5 μs|70.57 μs|3875.0000|0.0000|0.0000|15831.25 KB|
| **New** |	|	| **20,060.4 μs (-8%)** | **47.59 μs** | **3781.2500 (-2%)** | **0.0000** | **0.0000** | **15495.31 KB (-2%)** |
| Old |Push|100|13,718.6 μs|48.18 μs|640.6250|0.0000|0.0000|2679.69 KB|
| **New** |	|	| **11,058.1 μs (-19%)** | **40.45 μs** | **562.5000 (-12%)** | **0.0000** | **0.0000** | **2340.63 KB (-13%)** |
| Old |Index|100|15,416.2 μs|49.83 μs|828.1250|0.0000|0.0000|3422.66 KB|
| **New** |	|	| **12,151.2 μs (-21%)** | **27.03 μs** | **750.0000 (-9%)** | **0.0000** | **0.0000** | **3076.63 KB (-10%)** |
| Old |Map|100|4,999.5 μs|20.71 μs|1156.2500|0.0000|0.0000|4726.56 KB|
| **New** |	|	| **4,884.4 μs (-2%)** | **95.02 μs** | **1148.4375 (-1%)** | **0.0000** | **0.0000** | **4719.53 KB (0%)** |
| Old |Apply|100|782.5 μs|2.84 μs|215.8203|0.0000|0.0000|884.38 KB|
| **New** |	|	| **784.4 μs (0%)** | **2.88 μs** | **213.8672 (-1%)** | **0.0000** | **0.0000** | **877.34 KB (-1%)** |
| Old |JsonStringifyParse|100|4,908.0 μs|16.43 μs|1304.6875|0.0000|0.0000|5332.04 KB|
| **New** |	|	| **4,985.0 μs (+2%)** | **26.06 μs** | **1296.8750 (-1%)** | **0.0000** | **0.0000** | **5325 KB (0%)** |
| Old |FilterWithString|100|52,144.9 μs|139.18 μs|10700.0000|0.0000|0.0000|43886.1 KB|
| **New** |	|	| **53,303.0 μs (+2%)** | **79.17 μs** | **10700.0000 (0%)** | **0.0000** | **0.0000** | **43880.47 KB (0%)** |


## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|73.94 ms|0.279 ms|1571.4286|285.7143|0.0000|8614.83 KB|
| **New** |	|	| **73.71 ms (0%)** | **0.197 ms** | **1571.4286 (0%)** | **428.5714 (+50%)** | **0.0000** | **8602.27 KB (0%)** |
| Old |Run|dromaeo-core-eval|19.86 ms|0.055 ms|125.0000|0.0000|0.0000|523.86 KB|
| **New** |	|	| **19.91 ms (0%)** | **0.035 ms** | **125.0000 (0%)** | **0.0000** | **0.0000** | **518.22 KB (-1%)** |
| Old |Run|dromaeo-object-array|198.33 ms|0.802 ms|41000.0000|2000.0000|0.0000|175886.27 KB|
| **New** |	|	| **189.15 ms (-5%)** | **1.150 ms** | **41000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **175884.63 KB (0%)** |
| Old |Run|droma(...)egexp [21]|783.93 ms|7.975 ms|46000.0000|27000.0000|17000.0000|376754.9 KB|
| **New** |	|	| **749.28 ms (-4%)** | **11.060 ms** | **42000.0000 (-9%)** | **23000.0000 (-15%)** | **13000.0000 (-24%)** | **375524.06 KB (0%)** |
| Old |Run|droma(...)tring [21]|691.09 ms|6.751 ms|115000.0000|79000.0000|78000.0000|1349517.84 KB|
| **New** |	|	| **638.54 ms (-8%)** | **6.344 ms** | **111000.0000 (-3%)** | **76000.0000 (-4%)** | **75000.0000 (-4%)** | **1348409.46 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|159.41 ms|1.010 ms|2000.0000|0.0000|0.0000|8963.25 KB|
| **New** |	|	| **150.65 ms (-5%)** | **0.807 ms** | **2000.0000 (0%)** | **0.0000** | **0.0000** | **8798.48 KB (-2%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|317.7 ms|1.15 ms|49000.0000|2000.0000|0.0000|195.92 MB|
| **New** |	|	| **312.3 ms (-2%)** | **0.42 ms** | **49000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **195.92 MB (0%)** |
| Old |Benchmark|500|236.0 ms|0.23 ms|43000.0000|11000.0000|0.0000|173.43 MB|
| **New** |	|	| **240.9 ms (+2%)** | **0.62 ms** | **43000.0000 (0%)** | **11000.0000 (0%)** | **0.0000** | **173.43 MB (0%)** |